requirements.txt
pandas
numpy
tensorflow
spacy
streamlit
scikit-learn
pillow
git-lfs
-e .